Fix the logic for adding tasks
I realised that I don't want users to add tasks outside of projects. They must first create a project then add tasks to it. So I added a method called addTask to the Project class. This will also help the app to be more intuitive and guide the users to achieve a specific objective with every click on the app.